backend/GONXT backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
cryptography==41.0.7
Werkzeug==2.3.7
marshmallow==3.20.1
flask-cors==4.0.0
pytest==7.4.2
pytest-cov==4.1.0
requests==2.31.0
